Running: Cockpit is now in official Ubuntu backports
Drop the PPA install instructions. Backports are enabled by default on
all Ubuntu installations [1] so they are simpler to install, and are
available for all architectures.

Refer to documentation how to enable backports in case the user has
customized their apt sources to not include backports.

[1] with lower priority than the packages from the release, but as
Cockpit is entirely new in 16.04 and 16.10 this will just work.

Closes #67