Update ideals
Cockpit has been around for a while now and some of its ideals were out
of date or overlapping with others, some are simply features of the
software, and we developed a new one.

In "is discoverable", remove the sentence that says we're not there yet.
We're still not there, but much closer than we were when this was
written.

Remove "remotable". It's very obvious and even stated that it was.

Merge "does not own the server", "does not store data or policy", and
"reacts to the server" into "Cockpit directly interacts with and reacts
to the system". This removes a lot of overlap and brings down the number
of ideals.

Remove the bits about "pretty UI" from "zero footprint", because that's
just selling Cockpit short.

Remove "firewall friendly" and "pluggable", as those are established
features now.

Add usability testing to "cockpit is tested".

Also sort them differently. Roughly start with the ones that deal with
the qualities of Cockpit itself towards the qualities of making it.

Add a new one about being backwards-compatible and trying not to add new
APIs.