running: Update install instructions for RHEL
For RHEL ≥ 7.5, parts of Cockpit are in Base, but we still want to
recommend Extras to get cockpit-packagekit, and be able to install
cockpit-machines.

On RHEL 8, AppStream is enabled by default, and not having it is not
supported. So clarify that this first step only applies to 7.

On RHEL 8, the default zone has the Cockpit port opened in firewalld.
But point out that with custom zones it needs to be opened.

Fixes #242