data/browsers: Remove versions & rename Safari
Version strings are optional and are only used when our tests would show
a lower version. In other words, if we need to state a higher version,
we have an override. If the browser checks produce a higer version, that
will be used instead.

Safari has a manually higher version in place as it does not work
properly on macOS until this version, due to certificate issues.

iOS still has an issue with certificates, but there's not much more we
can do, aside from add a message about the issue and how a valid
certificate is 100% absolutely needed for iOS Safari (even if
a self-signed cert is fine for macOS Safari). This would be addressed by
a forthcoming update to the website.