running: Avoid partial upgrade in Arch installation instructions
`pacman -Sy <package>` is an unsupported operation that should be
avoided [1] as it can cause a partial upgrade of the system. Instead,
just document how to install only the cockpit package, and separately
document the (uncommon) case when repository indexes don't exist.

Thanks to NEXCydran for pointing this out!

[1] https://wiki.archlinux.org/index.php/System_maintenance#Partial_upgrades_are_unsupported

Fixes commit 36605790126
Fixes #313