Remove broken link to legacy Fedora packages services
The Fedora packages app has been deprecated for a while for now,
see https://discussion.fedoraproject.org/t/https-apps-fedoraproject-org-packages-is-down-and-has-been-for-a-while/23662/3?u=rugk
The suggested alternative is https://src.fedoraproject.org.

Unfortunately, I also found no way to link to the cockpit packages as before. There is only the usual page and also no way to show dependencies:
https://src.fedoraproject.org/rpms/cockpit

As such I just removed the link.