Add warning regarding removal of cockpit.css
I'm aware this post is known to be outdated and the starter kit should be used when getting into cockpit plugin development, but speaking from experience, this post is very enticing as a starting point (especially since going the route of [cockpit-project.org > `Contribute`](https://cockpit-project.org/external/wiki/Contributing.html) > Further Reading > `Create plugins for the user interface`, no mention of it being outdated is made). Anyone stumbling on to this page as a starting point will have a lot of frustration saved by including a note here on cockpit.css being removed.