applications: Add separate maintainers file
Should consolidate the maintainers file and make it a bit easier when it
comes to sorting and also not mess anything up when it comes to
third-party or new entries.

This also makes it easier to add new entries if we get other vendors
with many packages or noteworthy ones.

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>