_scripts: Fix errors and always raise exceptions
Few changes to fix a array access bug, import structure, image
versioning, and alt text generation. Also added an increment argument
instead of a "release" argument that was too vague.

Since the code takes from the main Cockpit repo this should not be an
issue and only be something that exists if you need it.

Changes:
- Import third-party dependencies at the bottom of the import-list
- Remove `##` headings from image alt text generation
- Use `version-name` structure for images
- Remove `--released` in favor of `--increment` to function in the
  opposite manner by incrementing instead of not incrementing.
- Add encoding to `open()`
- Don't silently catch exceptions during blog generation and instead
  raise it up the callstack.
- Moved class variable from global- to instance-based visibility

Signed-off-by: Freya Gustavsson <freya@venefilyn.se>