_scripts: revert arg for generate release notes
I had previously changed the `released` argument of generate release
notes script to function in reverse by renaming it to `increment`
instead.

My assumption had been that `released` was used when we didn't want to
update version for the blog post, and instead felt it would be easier to
have an `increment` argument instead.

This reverts it back to how it was before

Related-to: https://github.com/cockpit-project/cockpit-project.github.io/pull/837/commits/86c40fbed23b826b7750e95cc8ee326e0b0d770c
Signed-off-by: Freya Gustavsson <freya@venefilyn.se>